apply feedback from pr review
