+++ src/backend/fastapi_app/rag_simple.py
@@ -1,12 +1,9 @@
-import pathlib
-from abc import ABC, abstractmethod
 from collections.abc import AsyncGenerator
 from typing import Any
 
 from openai import AsyncAzureOpenAI, AsyncOpenAI, AsyncStream
 from openai.types.chat import ChatCompletion, ChatCompletionChunk, ChatCompletionMessageParam
 from openai_messages_token_helper import build_messages, get_token_limit
-from pydantic import BaseModel
 
 from fastapi_app.api_models import (
     AIChatRoles,
@@ -18,75 +15,7 @@ from fastapi_app.api_models import (
 )
 from fastapi_app.postgres_models import Item
 from fastapi_app.postgres_searcher import PostgresSearcher
-
-
-class ChatParams(BaseModel):
-    top: int = 3
-    temperature: float = 0.3
-    response_token_limit: int = 1024
-    enable_text_search: bool
-    enable_vector_search: bool
-    original_user_query: str
-    past_messages: list[ChatCompletionMessageParam]
-    prompt_template: str
-
-
-class RAGChatBase(ABC):
-    current_dir = pathlib.Path(__file__).parent
-    query_prompt_template = open(current_dir / "prompts/query.txt").read()
-    answer_prompt_template = open(current_dir / "prompts/answer.txt").read()
-
-    def get_params(self, messages: list[ChatCompletionMessageParam], overrides: dict[str, Any]) -> ChatParams:
-        top: int = overrides.get("top", 3)
-        temperature: float = overrides.get("temperature", 0.3)
-        response_token_limit = 1024
-        prompt_template = overrides.get("prompt_template") or self.answer_prompt_template
-
-        enable_text_search = overrides.get("retrieval_mode") in ["text", "hybrid", None]
-        enable_vector_search = overrides.get("retrieval_mode") in ["vectors", "hybrid", None]
-
-        original_user_query = messages[-1]["content"]
-        if not isinstance(original_user_query, str):
-            raise ValueError("The most recent message content must be a string.")
-        past_messages = messages[:-1]
-
-        return ChatParams(
-            top=top,
-            temperature=temperature,
-            response_token_limit=response_token_limit,
-            prompt_template=prompt_template,
-            enable_text_search=enable_text_search,
-            enable_vector_search=enable_vector_search,
-            original_user_query=original_user_query,
-            past_messages=past_messages,
-        )
-
-    @abstractmethod
-    async def run(
-        self,
-        messages: list[ChatCompletionMessageParam],
-        overrides: dict[str, Any] = {},
-    ) -> RetrievalResponse:
-        raise NotImplementedError
-
-    @abstractmethod
-    async def retrieve_and_build_context(
-        self,
-        chat_params: ChatParams,
-        *args,
-        **kwargs,
-    ) -> tuple[list[ChatCompletionMessageParam], list[Item]]:
-        raise NotImplementedError
-
-    @abstractmethod
-    async def run_stream(
-        self,
-        messages: list[ChatCompletionMessageParam],
-        overrides: dict[str, Any] = {},
-    ) -> AsyncGenerator[RetrievalResponseDelta, None]:
-        raise NotImplementedError
-        if False:
-            yield 0
+from fastapi_app.rag_base import ChatParams, RAGChatBase
 
 
 class SimpleRAGChat(RAGChatBase):
@@ -104,7 +33,7 @@ class SimpleRAGChat(RAGChatBase):
         self.chat_deployment = chat_deployment
         self.chat_token_limit = get_token_limit(chat_model, default_to_minimum=True)
 
-    async def retreive_and_build_context(
+    async def retrieve_and_build_context(
         self, chat_params: ChatParams
     ) -> tuple[list[ChatCompletionMessageParam], list[Item]]:
         """Retrieve relevant items from the database and build a context for the chat model."""
@@ -138,9 +67,7 @@ class SimpleRAGChat(RAGChatBase):
     ) -> RetrievalResponse:
         chat_params = self.get_params(messages, overrides)
 
-        # Retrieve relevant items from the database
-        # Generate a contextual and content specific answer using the search results and chat history
-        contextual_messages, results = await self.retreive_and_build_context(chat_params=chat_params)
+        contextual_messages, results = await self.retrieve_and_build_context(chat_params=chat_params)
 
         chat_completion_response: ChatCompletion = await self.openai_chat_client.chat.completions.create(
             # Azure OpenAI takes the deployment name as the model name
@@ -192,9 +119,7 @@ class SimpleRAGChat(RAGChatBase):
     ) -> AsyncGenerator[RetrievalResponseDelta, None]:
         chat_params = self.get_params(messages, overrides)
 
-        # Retrieve relevant items from the database
-        # Generate a contextual and content specific answer using the search results and chat history
-        contextual_messages, results = await self.retreive_and_build_context(chat_params=chat_params)
+        contextual_messages, results = await self.retrieve_and_build_context(chat_params=chat_params)
 
         chat_completion_async_stream: AsyncStream[
             ChatCompletionChunk
